src/requirements.txt
Django==3.1.7
redis==3.5.3
celery==5.0.5
django-environ==0.4.5
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
requests-html==0.10.0